conf: new function virDomainPCIAddressSetAllMulti()
This utility function looks through a virPCIAddressSet for any slot of
any bus that has multiple functions in use, and sets the "multi" flag
in the virDomainDeviceInfo for the device that is assigned to function
0 of that slot (as long as it hasn't already been set explicitly by
someone who presumably has better information than we do).

It isn't yet called from anywhere, so will have no functional effect.
Message-Id: <1484036583-6376-2-git-send-email-laine@laine.org>