Revert "storage: For FS pool check for properly formatted target volume"
The check does not work properly (crashes) with netfs filesystems and
also checking that a device is not empty when attempting to mount a
filesystem is not very usefull since the mount will fail anyways.

As the code would improve only a very minor corner case I don't really
see a reason to have this code at all.

This code would also fail if libvirt is compiled without support for
blkid and without parted.

This reverts commit a11fd69735e6951cda9bf256d8e423696a441aa4.
Message-Id: <d7b61fc441ebce5eba6608290f06fb82bd805442.1484230849.git.pkrempa@redhat.com>