qemu_domain: add timestamp in tainting of guests log
We lacked of timestamp in tainting of guests log,
which bring troubles for finding guest issues:
such as whether a guest powerdown caused by qemu-monitor-command
or others issues inside guests.
If we had timestamp in tainting of guests log,
it would be helpful when checking guest's logs(eg. /var/log/messages).

Signed-off-by: Chen Hanxiao <chenhanxiao@gmail.com>
Message-Id: <1484558350-5309-1-git-send-email-chen_han_xiao@126.com>