qemu: Actually unshare() iff running as root
While all the code that deals with qemu namespaces correctly
detects whether we are running as root (and turn into NO-OP for
qemu:///session) the actual unshare() call is not guarded with
such check. Therefore any attempt to start a domain under
qemu:///session shall fail as unshare() is reserved for root.

The fix consists of moving unshare() call (for which we have a
wrapper called virProcessSetupPrivateMountNS) into
qemuDomainBuildNamespace() where the proper check is performed.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <846142130b4d63f9ac3b299cd300cfa82d7103f2.1484652736.git.mprivozn@redhat.com>
Tested-by: Richard W.M. Jones <rjones@redhat.com>