conf: Introduce new hostdev device type mdev
A mediated device will be identified by the host PCI address of the
parent physical device which is backing the given mediated device, and a
UUID of the user pre-created mediated device. The data necessary to
identify a mediated device can be easily extended in the future, once we
need to enable managed='yes' in which case a hint from the upper
management layer about which mediated device type (e.g. vGPU type)
should an instance be created on.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <bb20b3ab4c80c895389d5e032630fd36a46642fd.1486383339.git.eskultet@redhat.com>