docs: Update RNG schema to reflect the new hostdev type mdev
To keep the domain XML as much platform agnostic as possible, do not
expose an element/attribute which would contain path directly to the
syfs filesystem which the mediated devices are build upon. Instead,
identify each mediated device by the parent physical device and a UUID
which is an optional element, but only if managed='yes' which is not
implemented yet.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <64b443867dac0e1915dfc7f6110e83dfa5643884.1486383339.git.eskultet@redhat.com>