qemu: Assign PCI addresses for mediated devices as well
So far, the official support is for x86_64 arch guests so unless a
different device API than vfio-pci is available let's assume PCI address
assignment. Once a different device API is introduced, we'll actually
have to start checking for the device API in the sysfs in order to
assign a correct type of address.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <e46b4a38a1acf957206f6a210285a30c9e7ca771.1486383339.git.eskultet@redhat.com>