qemu: Use nodedev callback mechanism to to get a nodedev data
Using the nodedev driver callback mechanism - allow qemu to be
notified whenever a nodedev is added/removed.

Keep track of the node devices in a hash table rather than requiring
a connection in order to get specific information about a node device
that qemu would eventually like to keep track of.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <1486611244-16608-3-git-send-email-jferlan@redhat.com>