qemu: Allow /dev/dri/render* for virgl domains
When enabling virgl, qemu opens /dev/dri/render*. So far, we are
not allowing that in devices cgroup nor creating the file in
domain's namespace and thus requiring users to set the paths in
qemu.conf. This, however, is suboptimal as it allows access to
ALL qemu processes even those which don't have virgl configured.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <1bebdb813af95cc777d07558096c2bf6599b3b7e.1486738487.git.mprivozn@redhat.com>