conf: Modify gendispatch.pl to make ACL argument opaque
The various drivers each have vir*EnsureACL and vir*CheckACL API's
which are generated on the fly by gendispatch to include the driver
typed "def" name as the last argument based on the API being checked
and the driver name (e.g. Node, Interface, Secrets, etc.).

Rather than having that name by typed, it's possible to generate the
API prototype using a "void *opaque" argument and then modify the first
line of the code to make the typed def argumment based off the opaque
data. This way we can generalize the ACL typedefs in object parsing
code to just be an opaque type. The next step in that generalization
is to remove the multiple typdef's and replace it with one general
typedef in order to further generalize the various drivers object
and object list handling functions.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <1486830585-23866-2-git-send-email-jferlan@redhat.com>