conf: Introduce virPoolObj and virPoolTableObj and PoolObj API's
Introduce a new generic virPoolObj[Ptr] and virPoolObjTable[Ptr] in order
to generically manage driver lists of objects rather than the current
hodgepodge of some drivers using a forward linked list and other drivers
using a single UUID only based hash table to look up objects.

The concept is to mimic the domain objects which keep objects in a UUID
hash table and a name hash table for faster lookup. There's also generic
API's in order to Lock, Reference, Unreference, and Unlock the objects to
remove the need for a heavy handed driver lock.

Since these are Hash Table objects rather than List Objects, the structures
and API's have been named thusly. These are all based upon the existing
generic object/class model so allocation/freeing is using existing API's.

In addition to common structures and API's, as each driver is converted
the driver specific ACL filter API's can be replaced by a generic pool
ACL filter API.

Each of the API's is designed to have a consistency with respect to returning
a locked and referenced object, so that a virPoolObjEndAPI can be called in
order to dereference and unlock the object.

The ultimate goal is to have a single way to think about these objects
management so that the various drivers don't have to have special or different
code in order to manage based on the "style" of list/table the driver needs
to manage.

A _virPoolObj uses a fairly generic structure:

    virObjectLockable parent;
    bool flags
    virPoolDefPtr pooldef;
    void *def;
    void *newDef;
    virFreeCallback defFreeFunc;
    void *privateData;
    void (*privateDataFreeFunc)(void *);

Access to the fields of the object is handled via API's to fetch or set
the various fields rather than allowing the calling driver to manage the
object directly. This ensures a more consistent look and feel to the code
for each driver.

The "def" and "newDef" field along with the defFreeFunc will point at
the driver's specific vir*DefPtr data and the FreeFunc handles the free
of the data when the object is finally removed.  The vir*DefPtr API's will
each need to be modified to take an "void *opaque" and assign that to the
specific driver definition type.

Most drivers have no need for privateData; however, a few drivers will
need to manage driver specific object private data. For those drivers
additional object management code is generated in order to manage the
Get, Set, and test of the private data. The goal being to have one set
of API's to manage the data rather than drivers directly touching it.

A _virPoolObjTable is another generic structure:

    virObjectLockable parent;
    virPoolObjTableType type;
    bool nameOnly;
    int hashStart;
    virHashTable *objsUuid;
    virHashTable *objsName;

Not every driver supports UUID's, so the UUID table isn't required, but
it is the preferred lookup mechanism.  The bool 'nameOnly' will help the
table code make those decisions.

The various Pool Object Table API's are taken from the current object list
managements and converted to use tables.

Objects are added to the table allowing for a callback mechanism to handle
driver specific "assignment" options when the object already appears in the
table and the driver would be managing the replacement of the "existing"
object "def".

There are common API's to Add, Clear, Remove, Search, Iterate, Collect,
List, Clone, and Prune objects. Each uses the same common methodology to
lock the table list, use the virHash* APIs, and unlock the table list
returning the requested data for the operation. Generally that means
a locked and referenced object. The virPoolObjEndAPI will handle the
dereference and unlocking.

Add new error code for the new poolobj subsystem

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <1486830585-23866-3-git-send-email-jferlan@redhat.com>